chore: Move GetScaledPath down to Uno.UWP (wasm)
